Финальный вариант лабы №4. Исправила моментики с типами врагов: их не было видно и работали мы с ними, как с обычными. Теперь враг с бронёй не пропускает фиксированное количество урона; враг с шансом уклонения иногда может пропустить весь наш урон; враг с шансом хила иногда может подлечиться сразу после нашего урона

+++ Game/CArmoredEnemyTemplate.cs
@@ -23,8 +23,8 @@ namespace Game
             armor = 25;
         }
 
-        public CArmoredEnemyTemplate(string name, string iconName, int baseLife, double lifeModifier,
-                                    int baseGold, double goldModifier, double spawnChance, double armor)
+        public CArmoredEnemyTemplate(string name, string iconName, int baseLife, int lifeModifier, // int
+                                    int baseGold, int goldModifier, int spawnChance, double armor) // int
             : base(name, iconName, baseLife, lifeModifier, baseGold, goldModifier, spawnChance)
         {
             Armor = armor;
